requirements.txt
beautifulsoup4==4.12.2
html5lib==1.1
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
playwright==1.41.2
prometheus-client==0.20.0
flasgger==0.9.7.1
PyJWT==2.8.0
requests==2.31.0
pytest==8.0.0
ruff==0.2.2
black==24.2.0
gunicorn==21.2.0
